Create documentation.md
because the world needs to know.

also a quick favor. if you would go to labels and make one called hacktoberfest-accepted and assign it to this pr it would be enough to make some great strides in the world of open source. Also mr gill survived the ventilator so hes kind of a fucking god and i will worship him for all my life.